Simplify the structure
